+++ cmd/util/prettier.js
@@ -3,23 +3,25 @@ const path = require('path');
 
 const argv = require('../../data/argv');
 
-const moduleName = argv.module;
-
 const projectConfig = require('../../project_config');
 
-let subDir = moduleName;
-if (moduleName === '*' || moduleName === 'all') {
-  subDir = '';
-} else if (moduleName.slice(-2) === '/*') {
-  subDir = moduleName.slice(0, -2);
-}
+const modules = argv.module.split(',');
+const moduleDirs = modules.map(moduleName => {
+  if (moduleName === '*' || moduleName === 'all') return '';
+  else if (moduleName.slice(-2) === '/*') return moduleName.slice(0, -2);
+  else return moduleName;
+});
 
 // Prettier needs to change `process.argv`.
 const oldProcessArgv = process.argv;
 process.argv = oldProcessArgv.slice(0, 2);
 
 process.argv.push(
-  path.join(projectConfig.buildPaths.src.dir, subDir, '**/*.{js,jsx,ts,css,less,sass,scss,json,md}'),
+  path.join(
+    projectConfig.buildPaths.src.dir,
+    moduleDirs.length < 2 ? moduleDirs[0] : `{${moduleDirs.join(',')}}`,
+    '**/*.{js,jsx,ts,css,less,sass,scss,json,md}'
+  ),
   '--write'
 );