+++ examples/dev/demo/project/src/test-5/index/js/a.js
@@ -1,10 +1,9 @@
-
 require('../css/index.css');
 
 const b = require('./b');
 
 module.exports = _ => {
-    console.log('test-5/index/js/a');
+  console.log('test-5/index/js/a');
 
-    b();
+  b();
 };