+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 0.4.0 [2018-07-26]
 
 - Remove `archive,arc,clean,revert` all server-side commands. 
-- Remove `ana` command. 
+- Remove `ana` command.
+- Fix bug of `eslint,stylelint,prettier` supporting `test/index,test-2/*` multi mode module. 
 
 ## 0.3.2 [2018-07-10]